Package the classes into 3 main categories
