+++ src/main/java/duke/data/TaskList.java
@@ -1,4 +1,7 @@
+package duke.data;
+
 import java.util.ArrayList;
+import duke.task.Task;
 
 public class TaskList {
