+++ src/main/java/duke/task/Deadline.java
@@ -1,3 +1,4 @@
+package duke.task;
 import java.time.LocalDate;
 import java.time.LocalTime;
