+++ src/main/java/duke/user/DukeException.java
@@ -1,3 +1,5 @@
+package duke.user;
+
 public class DukeException extends Exception {
     public DukeException () {
         super(" ☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
